@@ -5,17 +5,17 @@
   "proxy": "http://localhost:3001",
   "dependencies": {
     "@apollo/client": "^3.5.8",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
     "graphql": "^15.4.0",
     "bootstrap": "^4.4.1",
-    "jwt-decode": "^2.2.0",
-    "react": "^16.13.1",
+    "jwt-decode": "^3.1.2",
+    "react": "^17.0.1",
     "react-bootstrap": "^1.0.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1"
+    "react-dom": "^17.0.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "4.0.1"
 
   },
   "scripts": {
