@@ -2,24 +2,24 @@ const { gql } = require('apollo-server-express');
 
 const typeDefs = gql `
 type User {
-    _id: ID
-    username: String
-    email: String
-    bookCount: Int
+    _id: ID!
+    username: String!
+    email: String!
+    bookCount: Int!
     savedBooks: [Book]
 }
 
 type Book {
-    bookId: ID
+    bookId: ID!
     authors: [String]
     description: String
     title: String
-    image: 
+    image: String
     link: String
 }
 
 input bookInput {
-    bookId: String
+    bookId: String!
     authors: [String]
     description: String
     title: String
@@ -39,11 +39,9 @@ type Auth {
   type Mutation {
     login(email: String!, password: String!): Auth
     addUser(username: String!, email: String!, password: String!): Auth
-    saveBook(input: bookInput): User
+    saveBook(input: bookInput!): User
     removeBook(bookId: String!): User
   }
-
-
 `;
 
 module.exports = typeDefs 
